refactor: merge www/ and root
