+++ data/version.json
@@ -1,0 +1,5 @@
+{
+  "file": {
+    "version": "1"
+  }
+}
