+++ routes/[id]/[file].ts
