--- www/deno.json
@@ -1,22 +1,0 @@
-{
-  "tasks": {
-    "start": "deno run -A --watch=static/,routes/ dev.ts",
-    "build": "deno run -A dev.ts build",
-    "preview": "deno run -A main.ts"
-  },
-  "compilerOptions": {
-    "jsx": "react-jsx",
-    "jsxImportSource": "preact"
-  },
-  "importMap": "./import_map.json",
-  "lock": false,
-  "lint": {
-    "rules": {
-      "tags": ["fresh", "recommended"]
-    },
-    "exclude": ["_fresh"]
-  },
-  "fmt": {
-    "exclude": ["_fresh"]
-  }
-}
